code_challenge/pyproject.toml
[tool.poetry]
name = "code-challenge"
version = "0.1.0"
description = ""
authors = ["Robert Tsui <rt.roberttsui@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
